requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
dataclasses==0.6
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
importlib-metadata==1.3.0
Jinja2==2.10.3
kombu==4.6.7
MarkupSafe==1.1.1
more-itertools==8.0.2
psycopg2-binary==2.8.4
PyJWT==1.7.1
pymongo==3.9.0
pytz==2019.3
redis==3.3.11
six==1.12.0
sqlparse==0.2.4
vine==1.3.0
zipp==0.6.0
